fixed some null refs and added in some more logic
i added in the call to the timer to get the completed bool to end the game. critter can now flip left and right. the logic for keelys level is done for the eqn. still need to work on the menu functionality, its a little weird to control  it with keys.